Add more detail to README.md and macOS support
Add more detail in the README with the compiled binary syntax, examples of the binary being used, footnotes regarding macOS compilation and dependencies, and a macOS dcraw binary. Regarding the macOS dcraw binary, I could not get the included binary working under macOS so I copied the binary provided in the brew package "dcraw" and included it here with instructions on how to change what dcraw binary is refrenced.